pyproject.toml
[project]
name = "baseline-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "einops>=0.8.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pyyaml>=6.0.3",
    "rasterio>=1.5.0",
    "scikit-image>=0.26.0",
    "scipy>=1.17.1",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
]

[tool.uv]
index-strategy = "unsafe-best-match"

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
